+++ lib/lastfm_archive/behaviour/transformer.ex
@@ -3,8 +3,11 @@ defmodule LastfmArchive.Behaviour.Transformer do
   Transform strategy for applying post archive side affects
   """
 
+  @type data_frame :: Explorer.DataFrame.t()
   @type metadata :: LastfmArchive.Archive.Metadata.t()
   @type options :: keyword()
 
-  @callback apply(metadata, options) :: {:ok, metadata()} | {:error, term()}
+  @callback source(metadata(), options()) :: data_frame()
+  @callback transform(data_frame()) :: data_frame()
+  @callback sink(data_frame(), metadata(), options()) :: :ok
 end